Fixed bugs in make-worksheets.sh, based on feedback from Eric Loots.

+++ make-worksheets.sh
@@ -1,22 +1,36 @@
 #!/usr/bin/env bash
 
 script=$0
-default_target_dir=src/worksheet
+default_source_dir=src/script/scala
+default_target_dir=src/main/scala/worksheet
 default_ext="worksheet.sc"
+
+
+fatal_message() {
+cat <<EOF
+NOTE: If worksheet evaluation fails, try removing the -Xfatal-warnings flag
+in build.sbt.
+EOF
+}
 
 help() {
   cat <<EOF
-Copies src/script to src/worksheets and renames all the files to "*.worksheet.sc",
-the default extension for VSCode worksheets.
+Copies $default_source_dir to $default_target_dir, the default location for Metals,
+and renames all the files to "*.worksheet.sc", the default extension for
+VSCode worksheets.
 
-usage: $script [-h|--help] [-t|--target dir] [-e|--ext extension]
-where:
+$(fatal_message)
+
+Usage: $script [options]
+
+Where the options are the following:
 -h | --help           Shows this help and quits
 -f | --force          Force deletion of old $target_dir location.
                       By default, it won't overwrite an old copy.
+-s | --source dir     Read the scripts from "dir" (default: "$default_source_dir").
 -t | --target dir     Write results to "dir" (default: "$default_target_dir").
 -e | --ext extension  Use ".extension" (default: ".$default_ext").
-                      (If you prepend a ".", it will be handled correctly.)
+                      (A leading "." is optional; it will be added if needed.)
 EOF
 }
 
@@ -27,6 +41,7 @@ error() {
   exit 1
 }
 
+source_dir=$default_source_dir
 target_dir=$default_target_dir
 ext1=$default_ext
 force_delete=
@@ -43,6 +58,10 @@ do
     -t|--t*)
       shift
       target_dir=$1
+      ;;
+    -s|--s*)
+      shift
+      source_dir=$1
       ;;
     -e|--e*)
       shift
@@ -67,9 +86,21 @@ fi
 
 ext=$(echo $ext1 | sed -e 's/^\.*//')
 
-cp -r src/script $target_dir
-find $target_dir -name '*.scala' | while read f
-do
-  ff=${f%.scala}
-  $NOOP mv $f $ff.$ext
-done
+if [[ -z $NOOP ]]
+then
+  cp -r $source_dir $target_dir
+  find $target_dir -name '*.scala' | while read f
+  do
+    ff=${f%.scala}
+    $NOOP mv $f $ff.$ext
+  done
+fi
+
+cat <<EOF
+
+Finished copying $source_dir to $target_dir and changing file extensions
+to $ext.
+
+$(fatal_message)
+
+EOF
